Refactoring into 'src' package
