+++ src/tw/DistanceCalculator.py
@@ -4,9 +4,9 @@
 
 Takes in the entire list of villages and generates stats of distance and attack time
 """
-from core.Villa import FarmVilla as Villa
-from core.VillaStats import Stats
-from util.Helper import read_generic_config
+from src.core import FarmVilla as Villa
+from src.core import Stats
+from src.util.Helper import read_generic_config
 
 
 class DistanceCalc:
