+++ src/tw/ExtractBarbs.py
@@ -10,9 +10,9 @@ import requests
 from bs4 import BeautifulSoup
 from ordered_set import OrderedSet
 
-from core.Villa import FarmVilla as Villa
-from tw.ProcessLocalConfig import WorkerProcessor
-from util.Helper import read_config_world_level
+from src.core import FarmVilla as Villa
+from src.tw.ProcessLocalConfig import WorkerProcessor
+from src.util.Helper import read_config_world_level
 
 
 class _ExtractBarbsList:
