+++ src/tw/ProcessLocalConfig.py
@@ -9,9 +9,9 @@ import os
 from jinja2 import Template
 from ordered_set import OrderedSet
 
-from core.Villa import FarmVilla as Villa
-from tw.Template import local_config_template
-from util.Helper import read_json_obj_world_level, write_json_to_fileSystem, get_current_time
+from src.core import FarmVilla as Villa
+from src.tw.Template import local_config_template
+from src.util.Helper import read_json_obj_world_level, write_json_to_fileSystem, get_current_time
 
 
 class WorkerProcessor:
